Practica 1/backend/requirements.txt
flask
psycopg2-binary
pytest
pytest-cov
coverage
black
flake8